Fix: get replays working again (#5776)
Fix the defect in the existing replay code that I made when I switched from
use a `QTime` to a `QElapsedTimer` in PR #4400. I neglected to account for
the elapsed() method's return type changing from an `int` (effectively a
`int32_t`) to a `qint64` (a.k.a. `int64_t`). This PR attempts to get around
the issue by trying to read the two integer values that precede each
"chunk" of replay data in the recording first as two 4 byte `int32_t`s
(actually `qint32s`) and if that does not work (because it produces
obviously wrong values - negative or overlarge) to try again with the first
of the values as an 8 byte `qint64`.

This commit also corrects the code that produces a recording file to use a
4 byte `qint32` as the first value, restoring the format to that used prior
to the altered form in the original PR - which was introduced after
Mudlet 4.10.1 was released.

This should close #5770.

This includes raising 2 issues: #5779 and #5780 to handle new TODO:
comments.

Release post highlight
================
Fix: repair replay functionality to work again - and prevent Mudlet
crashing when trying to replay a recording made in Mudlet versions after
Mudlet 4.10.1. Replay made in the intervening period will be safe to load
in Mudlet versions after this fix is introduced; also replays made after
this will also be safe to load in Mudlet versions of that period.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>