Improve: add flag and getter to track whether MSDP has been negotiated
I am extending the statistics reporting (from the button in the editor)
and was planning to display some extra tables besides the ATCP, GMCP and
Channel 102 ones. I wanted to only display each table if it was in use
and it seems there is not an `(bool) cTelent::isXxxxxEnabled()` getter
for the MSDP OOB data channel.

As the prototype PR I had put together hit just more than 10 files I
realised that I could spin this aspect out to a small precusor one for
ease of reviewing.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>