Improve: clean-up/unify and extend item statistics reporting system
Steps taken:
* removes unwise '_' prefix on `(void) _assembleReport(T*)` methods
* rename `(void) XxxxUnit::initStats()` to `(void) XxxxUnit::resetStats()`
or provide it in classes without it - and use it in top-level
`(std::tuple<QString, int, ...int>) XxxxUnit::assembleReport()` methods
* remove call to aforesaid method in `XxxxUnit` constructors - as that can
be done more simply/cleanly as part of initialisation in the header.
* rename alias/key/timer/... etc elements in `XxxxUnit::assembleReport(...)`
methods to be type neutral `item`/`pItem`/`pChild` - this reduces the
typographical differences in the methods for each type - and may make it
possible to abstract them to a template class later on (for **DRY** reasons!)
* after some code rearrangement it became clear that part of the:
`XxxxUnit::assembleReport()` top level instance code was a recursion of
the ``XxxxUnit::assembleReport(...)` code - so could
replaced by a call to that for more compact code!
* make the stats counters in each `XxxxUnit` private and remove the unused
ones - this typically leaves only three in each class.
* refactor the reporting to the `TMainConsole` code more compact AND
redesign it for I18n/L10n as it is UI text presented in the main console so
should be presented in a manner appropriate to the user's locale setting.
* extend the reported statistics to include Scripts and Timers - which were
previously missing - this addresses the comments around:
https://github.com/Mudlet/Mudlet/pull/5706#discussion_r758632670
* extend the reporting to include other tables include MSDP, this will
close #692 (specifically it adds the ones for MSSP and MSDP and if there
are any others that need to be added they can probably be included with a
couple of lines of code).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>