Add missing cursor and selection functionality to Geyser.MiniConsole (#5851)
#### Brief overview of PR changes/additions

Someone asked how to get the first line in a miniconsole in Lua, and I realized as I was answering this was a place that Geyser was missing functionality it should have. This PR adds the functions for moving and selecting text with the virtual cursor, and getting the information under the selection.

It also adds in a couple of get* functions which were missing, but I can see the argument for pulling them into their own PR. I ran across them while searching for the cursor related functions but they aren't strictly related.

#### Motivation for adding to Mudlet

API consistency and completeness.

Should not have to drop out of the OO paradigm of Geyser and use the base Mudlet functions the objects wrap for manipulating the cursor/selecting text. Ideally, we wrap anything that needs to have geyserObject.name passed into it to return the result specific to that item.

#### Other info (issues closed, discussion etc)

This will catch UserWindows as well, as they inherit from Geyser.MiniConsole

#### Release post highlight
<!--
Use this space if you wish to write a short statement or example for inclusion
in the release post for the next release. 
-->