Improve: fixup layout and tab-ordering for preferences dialogue (#5846)
The prior code erroneously had a `QLayout` as a display element (for the empty part of the preferences which is populated with the main window shortcuts when the dialogue is created. Whilst the C++ code to add the details for the shortcuts to that `QLayout` is correct, the layout should be attached to a parent `QWidget` derived item itself. (It was showing up in RED indicating a problem, in the Qt Designer tool/utility).

As well as fixing that by adding a `QFrame` widget for the layout, this code also converts a higher up layout from a `QGridLayout` to the easier to maintain `QVBoxLayout` and assigns some more meaningful names to some layout items, so that they can be associated with their parent widget which makes things easier when editing the file by hand...

Also fixes up the tab ordering for this and some other recently added widgets.

Also fixes up the disabling and enabling some recently added Host specific options so they are not accessible when there is not a profile loaded.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

#### Release post highlight
Fix: correct some controls being in the wrong tab order and not being disabled when there is not a profile loaded for them to work on in the profile preferences dialogue.