Fix: prevent bogus NAWS and window resize events on tab changes
This should prevent NAWS messages and `sysWindowResizeEvents` with a width
of 0 being raised when the user is multi-playing but not using multi-view
and changes to a different tab. As this also stops that width of zero size
being stored it also means a second resize event/message when the user
switches back to the profile concerned is also not produced.

It is possible that these bogus messages of zero width might have been used
to detect when the particular profile lost the focus when multi-playing but
not using multiview - as such there might be a desireable for a new event
to report when a profile loses/gains the state of being the "active"
profile...

This should close #5814.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>