Improve: (2D) mapper - fade out exit lines to out of viewport rooms
I conceived of this without being aware that the change in part of #4608
that stopped drawing the entire exit line between two rooms both visible in
the mapper twice (once for each room) had been reverted by #5949.

This is a suggested alternative that goes back to drawing the half a line
from each room towards the other room but additionally draws the other
section stating as fully opaque but fading to transparent at the far room.
If both rooms are visible this will be hidden by the corresponding fully
visible exit line from the other room - but not if not.

It is a bit more complex to do than might be first thought because the
`QPainter::drawLine(...)` methods only use the `QPainter`'s pen (otherwise
used to draw the outline of a shape) and a gradient shading effect only
applies to the brush (fill) of a shape. This means that the line has to be
constructed as a rectangle - and to draw it at an angle it needs to be done
with `QPainter::drawPolyon(...)` using a `QLinearGradient` as the brush and
no pen.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>