Revise: adopt peer-review suggestions
Small arrows are now the default. The user will have to ask to use the
large ones - which means that existing maps should not have an issue with
the area exit arrows now - unlike as has been reported for the 4.15.0
release version.

A scaling factor used with different values for small and large area exit
arrows has been changed to be only generated once per call to
`(void) T2DMap::paintRoomExits(...)` instead of once or twice depending on
the direction of the exit in the XY-plane.

A tooltip for this *knob* has been removed as it specifically mentioned the
version of Mudlet that introduced the large area exits - and it doesn't
need to now the default is pretty much as Mudlet would have been before
that version.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>