Fix: add licenses for bundled luarocks modules
When we make an installable or distributable build we compile some files
from various Luarocks and as they are under the same license as Lua (MIT)
we MUST mention them and their license in the 3rd party components tab in
the "About Mudlet" dialog for those builds that we distribute from the
https://www.mudlet.org site. However users building Mudlet personally are
more likely to source the luarocks themselves and as such it is not
relevant to mention these things. Setting the environmental variable
`WITH_BUNDLED_LUA_FILES` to `NO` will, as a result of this PR suppress
mention of them from the 3rd party component tab but otherwise they will
show as required.

I have also taken the oportunity to clean up some other details in the
"3rd Party" tab in the "About Mudlet" dialogue - many some initial capital
letters and formatting of some included URLS. Also a typo of `<ht>` instead
of `<hl>`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>