Fix: revise/improve new map_label form/dialogue
Reorder the elements in a `QGridLayout` so that they are in a logical
order.

Convert a `QGroupBox` to a `QFrame` and move it's title element to the
whole "dialogue" (which otherwise would display the default "Form" to the
end-user!)

Add buddy definitions to all the labels to their corresponding control
widgets.

Give more meaningful (object) names to some objects that otherwise have the
Qt defaults of:
`widgetClassWithoutLeadingQ(_numeric_suffix_if_more_than_one)`.

Consideration ought to be given to converting the lower 2 push-button
composite widget to be the more usual `QDialogButtonBox` that is used in
this position...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>