Fix: switch to 5.1-syntax_5.1 branch for lcf submodule
For the last few years we have been using the wrong branch as the one we
were on did not permit the use of `goto` as an identifier and "labels" as
targets for such `goto`s as version of Lua after 5.1 forbode that as it
became a command to uncontrollably jump around Lua scripts as it does in
the C (or BASIC) programming languages!

This change will pull in changes that @keneanung did to the lfc submodule
as:
* https://github.com/martin-eden/lua_code_formatter/pull/15
* https://github.com/martin-eden/lua_code_formatter/pull/16
but we had forgotten to use.

This will close #5997.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>