Fix: correct the QMake variable used to record some project files
In the previous commit I added some files to the `DISTFILES` qmake variable
but that is for files that should be included in a `make dist` (source code
tarball) Unix makefile usage. Given that the `.github` directory files are
not ones that would be used in this manner (they are only used by the
Mudlet project to run it's own CI/CB process) they should instead be
assigned to the (undocumented in the qmake documentation) `OTHER_FILES`
variable.

There are further changes needed to the QMake `OTHER_FILES`, `INSTALLS` and
`DISTFILES` variables but they should be left to a further, separate PR.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>