Revise: include the licenses for the Luarock modules even when NOT required
This was requested during peer-review and it doesn't make sense to me - why
report about items that are not inherently in the source code and which
we do not distribute with that - when the only reason we do have to mention
them is when we ship their binaries as part of a precompiled build? OTOH
it has been pointed out that it does remove one "option" that self-buildera
could frob when they build Mudlet from source so will make things
marginally simpler for them.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>