Add: new function moveRoom(...)
This can move a room AND it's custom exit lines
`setRoomCoordinates(...)` does NOT move the latter, for that
the custom line stays the same except the start position of
the *first* segment moves with the room.

There are multiple argument options summarised as:
* `moveRoom(roomId, deltaX, deltaY)` -
shift the given room and its custom exit lines by the given
amounts in the x and y directions.
* `moveRoom(roomId, deltaX, deltaY, false)`
- as above.
* `moveRoom(roomId, coordX, coordY, true)`
- move the room and its custom exit lines to the given
coordinates on the same (z) level.
* `moveRoom(roomId, deltaX, deltaY, deltaZ)` - shift the given
room by the given amounts in the x, y AND z directions, any
custom exit lines will also move up or down to stay on the
same level as the room.
* `moveRoom(areaId or areaName, roomId, deltaX, deltaY, deltaZ,
                                                        false)`
- as above.
* `moveRoom(areaId or areaName, roomId, coordX, coordY, coordZ,
                                                         true)`
- move the room to the given coordinates, any custom exit
lines for the room will also move to the given z coordinate.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>