Add: new function moveMapLabel(...)
This is similar to the new `moveRoom(...)` in PR #6010 and it
provides the ability to move a map label either to an
absolute position (for the top-left corner) or relative to its
current position in a similar manner.

There are multiple argument options summarised as:
* `moveMapLabel(areaId/Name, labelId/Text, deltaX, deltaY)` -
shift the label identified by the given zero or greater ID
number or the first label with the given text in the area with
the given ID number or name by the given amounts in the x and
y directions; x and y are floating point numbers and they are
in the same coordinates as the map rooms. I.e. moving a label
by 1.00f in a direction moves it the same amount as a room
that is moved by 1 in the same manner. The label's z
coordinate is not changed.
* `moveMapLabel(areaId/Name, labelId/Text, deltaX, deltaY,
                                                      false)`-
as above.
* `moveMapLabel(areaId/Name, labelId/Text, deltaX, deltaY,
                                                       true)`-
move the top-left corner of the map label room to the give
x and y coordinates, the level (z coordinate) on which it is
shown is unchanged.
* `moveMapLabel(areaId/Name, labelId/Text, deltaX, deltaY,
                                                    deltaZ)` -
shift the specified label by the given amounts in the x, y AND
z directions, although the z value may be a floating point
number it is rounded to an integer equivalent.
* `moveMapLabel(areaId/Name, labelId/Text, deltaX, deltaY,
                                             deltaZ, false)` -
as above.
* `moveMapLabel(areaId/Name, labelId/Text, deltaX, deltaY,
                                              deltaZ, true)` -
move the top-left corner of the map label room to the give
x and y coordinates, the level (z coordinate) on which it is
shown is also changed though it may be a floating point number
it is rounded to an integer equivalent.

To detect whether a label has been found given that the
container that stores them in a `TArea` instance stores them
directly (instead of using pointers to `TMapLabel` instances
keep on the heap) I've added an `isNull()` method to that
class which is used to check whether a sought for instance got
found and copied to a local `automatic` storage instance.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>