Revise: move changes to new getCustomLines1(...)
This reverts the change made to `getCustomLines(...)` and instead
produces a new function with the fix applied. So as to make it
easier to pass the returned data back into `addCustomLine(...)`
(it can also be used to replace/update an existing line) the output
format for the color and the coordinates have been reformatted so
the tables can be sent as they are to the latter function. This
means that instead of the entries for the color being keyed against
`"r"`, `"g"` and `"b"` they are instead reported as a *list* i.e.
with keys of `1`, `2` and `3`. Similarly for the coordinates of the
points of the custom line - instead of `"x"` and `"y"` being used
they are keyed to `1` and `2` respectively along with a third `3`
key for the Z-axis coordinate which is the same as that of the
room, for EVERY point - which is what the `addCustomLine(...)`
function uses - along with each point being indexed with a number
now starting at the Lua standard of `1` instead of the previous
`0`.

This change is so that we do not break backwards compatibility with
anything (and one product has already been identified) that used
the previous function but which likely had a external fix being
done on the incorrectly indexed points.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>