Add: new function getCharacterName()
This is one portion of the stalled PR #3952 that can be lifted from there
and be implemented without raising any eyebrows (IMHO).

It returns the string that the user entered into the "Character name:"
field on the "Connection profile" dialogue and is needed to allow a shared
module to customise its behaviour depending on the particular profile that
it is being used with on the basis of the identifier being sent for login
purposes - *c.f.*
[`getProfileName()`](https://wiki.mudlet.org/w/Manual:Lua_Functions#getProfileName)
which instead identifies the profile name which is very likely not the same
text.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>