Add: new function getCustomLines1(...) to fix ordering of custom line points in original getCustomLines(...) (#6009)
Originally titled:
"Fix: ordering of custom line points in getCustomLines(...)"

Make the first point be at an index of 1 (which is standard
for Lua) instead of 0 (C/C++ array standard) so that the table
of points return can be passed back to addCustomLine(...)
without having to fiddle with a zero-th element which would be
out of order otherwise.

This was discovered last night whilst I was looking at code
to move rooms AND their associated custom exit lines using the
Lua API. As it is up to now a fix is needed which can be
achieved by a function that @demonic came up with:

```lua
function fix0index(tbl)
  local result = {}
  if #tbl ~= table.size(tbl) - 1 then
    debugc("This is not a list table, cannot fix the 0 index")
  end
  if not tbl[0] then
    return tbl
  end
  for i = 0, #tbl do
    result[i+1] = tbl[i]
  end
  return result
end
```

then changed to reverts the change made to `getCustomLines(...)`
and instead produce a new function with the fix applied. So as to
make it easier to pass the returned data back into `addCustomLine(...)`
(it can also be used to replace/update an existing line) the output
format for the color and the coordinates have been reformatted so
the tables can be sent as they are to the latter function. This
means that instead of the entries for the color being keyed against
`"r"`, `"g"` and `"b"` they are instead reported as a *list* i.e.
with keys of `1`, `2` and `3`. Similarly for the coordinates of the
points of the custom line - instead of `"x"` and `"y"` being used
they are keyed to `1` and `2` respectively along with a third `3`
key for the Z-axis coordinate which is the same as that of the
room, for EVERY point - which is what the `addCustomLine(...)`
function uses - along with each point being indexed with a number
now starting at the Lua standard of `1` instead of the previous
`0`.

This change is so that we do not break backwards compatibility with
anything (and one product has already been identified) that used
the previous function but which likely had a external fix being
done on the incorrectly indexed points.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>