Revise: swap some files between QMake OTHER_FILES and DISTFILES variables
Technically the latter are for files that are needed to go into a source
tarball that the traditional GNU `make dist` operation would produce. As
such it does not seem reasonable to include CI/CB control files in it.
Conversely the run-time LUA files that Mudlet needs definetely should be
included and so this commit moves them from the `OTHER_FILES` variable,
whos contents are merely files that we want to show up in the Qt Creator
IDE so they can be seen, searched and edited from there.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>