Cleanup: undo unwanted changes from previous commits
Now that a solution seems to have been found it is time to clean up the PR.
One unneed change can be left in place - is the elimination of some
auxillary parts of libzip. Note that the failure to build in the previous
commit was due to an addition that this commit is removing (trying to force
make to only run a single job at a time). So this should be the required
fix.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>