Infrastructure: repair Windows CI build failures (#6072)
During running of the CI builds on AppVeyor for Windows we now are getting errors during the `.configure` stage for the build of PCRE - and we get the advices to try using the additional `--disable-dependency-tracking` to fix it. This PR tried to do that to see if it fixed things, but it didn't.

Further investigation by @keneanung and myself eventually revealed that although #6030 installed Qt it was not installing the expected Mingw GCC compiler (32-bit 7.3.0) - further digging revealed that an addition step was needed to do that.

During the investigation I spotted that when building `libzip` we were also building the regression tests, the documentation and the examples - none of which are used or needed by our Windows CI builds. So disabling them should reduce the build time slightly and there was no point to undoing the change that disabled their being built.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>