Fix: return empty table if no stub-exits for getExitStubs functions
The existing code produces a nil+warning message if there are no exit stubs
for a room when `getExitStubs(...)` OR `getExitStubs1(...)` are called on
it. This is unfortunate because the only way to find out whether a room has
any exit stubs in a script is by using those functions.

Unlike the draft #6069 this does not produce a `nil` *without* a message
in just this "problem" case, instead it just returns an empty table, so the
table can be stored locally and examined in both the case where there are
and are not exit stubs. I feel this is a cleaner solution because there
are other run-time problem situations that also will produce a `nil` and
a message - which that draft doesn't really resolve IMHO.

To enable this fix (and thus to change the behaviour away from how a
previous script/package might be working to examine the precise error
message and thus not be entirely backward compatible) it is necessary to
set the global boolean variable `noWarningForNoStubExits` to `true` - this
has the nice side-effect that a script/package could test for the existance
of this variable to test whether the fix is available!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>