Fix: restore Website label visibility on Connect Profile dialogue
This will close #6087 and address one task in #6039. The latter also
involved renaming the widget for the website label from `label_7` to
`label_website`.

It turns out that there is no margin/space set around the elements of the
informational area, so it was necessary to add an "indent" to the "Website"
label to push it away from the left side. Also, by changing the alignment
of the details that are shown for the website for the selected MUD to be
centered in both horizontal AND vertical direction it removed the need
for each entry in the:
`(OrderedMap<QString, GameDetails>) mudlet::scmDefaultGames`
that populates that field to include the HTML-like `<center>`...`</center>`
tags to produce the same (horizontal centering) effect.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>