Infrastructure: correctly #include winsock2.h for Windows builds
One or more Qt libraries happens to cause the `windows.h` header file
to be `#include`d; unfortunately this will also pull in the original
`winsock.h` header file which is contains some definitions that clash with
those from the (preferred) `winsock2.h` header file. The best way to
prevent getting warnings of:
"#warning Please include winsock2.h before windows.h [-Wcpp]"
is to include the latter file where it is needed, *before* a Qt library
can pull in the older (and obsolete) header.

As `Q_OS_WIN32` is actualy defined in a Qt header file it is not suitable
to use **that** being defined as the test for the inclusion and instead
I have arranged for `INCLUDE_WINSOCK2` to be defined by the (QMake and
CMake) build systems' project file to trigger the required `#incluse`s
for **Windows** bulds.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>