Infrastructure: remove hidden 'Website:' label from connection dialog (#6111)
This will close #6087 and address one task in #6031 

The original plan was to reposition and show this label but It turns
out that we don't really want it anyhow. With it's removal there is no
need to retain the remaining two widgets in a `QGridLayout`, and
they can be placed in an easier to maintain `QVBoxLayout` instead!

Also we will hide the label that shows the Website details when there
are none to display. This was briefly discussed between me and
Vadim on Discord and felt to be a good idea - as it leaves more space
for useful other content (the MUD description or user entered notes).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>