Revise: getExit{|1}(...) return empty table unconditionally on no stubs
This changes from previous behaviour of a `nil` plus an error message when
there were no stubs to report and of just returning a `nil` in even older
versions of Mudlet. Technically this could be held to be breaking backwards
behaviour but was suggested as desirable by both @keneanung and @vadi2.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>