Fix: return empty table if no stub-exits for getExitStubs functions (#6081)
The existing code produced a nil+warning message if there are no exit stubs
for a room when `getExitStubs(...)` OR `getExitStubs1(...)` are called on
it. This is unfortunate because the only way to find out whether a room has
any exit stubs in a script is by using those functions and the error message
is not really helpful then!

Unlike the draft #6069 this does not produce a `nil` *without* a message
in just this "problem" case, instead it just returns an empty table, so the
table can be stored locally and examined in both the case where there are
and are not exit stubs. I feel this is a cleaner solution because there
are other run-time problem situations that also will produce a `nil` and
a message - which that draft doesn't really resolve IMHO.

This change from the previous behaviour or that of just returning a `nil` in
even older versions of Mudlet could technically be held to be breaking
backwards behaviour but was suggested as desirable by both
@keneanung and @vadi2.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>