Fix: use native language for desciptions of a couple of predefined MUDS
Following the inclusion of "Mud.ren" with its Chinese description text
(in https://github.com/Mudlet/Mudlet/pull/6140),
I felt it was appropriate to put in the native language texts for a
Spanish and an Italian MUD who have already provided us with the text
to use but which we hadn't so far. This will address points which I
made about this in:
https://github.com/Mudlet/Mudlet/pull/1505
and
https://github.com/Mudlet/Mudlet/pull/2502#discussion_r273176458
respectively...

I have cleaned up the texts involved for all the descriptions so
that they are all wrapped in our `qsl(`...`)` wrapper and the
host URLs that are used to identify which one to use are all wrapped
in `QLatin1String(`...`)` as they should all be ASCII (as they are
all OS compatible file names as well) and there is a
`QLatin1String::operator==()` method for them.

In passing I have also "fixed" the "website" entries for the Spanish
MUD as it is appropriate for them to be in Spanish rather than English
as they are not URLs but HTTP-like `<a href="xxxx">`...``</a>` entries.

I propose to follow this PR with another one to move the description
data to the same place as all the other data about the predefined MUDs
(the container defined in the `./src/mudlet.h` file:
`(inline static const OrderedMap<QString, GameDetails>)
                                               mudlet::scmDefaultGames`)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>