Fix: ensure liboleaut32 library is used for both QMake Windows Setups
#6090 added the library to the QMake LIBS variable for the main build
system but it is in fact needed also for the MSYS2/Mingw-w64 setup (that I
use) as well so needs to be put in a different location in the main QMake
project file.

Whilst checking to see if the CMake alternative worked in my setup I
noticed that there is a typo in the adjecent line for the "ws2_32" library.
The documentation at, say:
https://docs.microsoft.com/en-us/windows/win32/winsock/initialization-2
calls it `Ws2_32`, however the files in a Windows filesystem (and the
corresponding line in the CMake build file):
```
if(WIN32)
  target_link_libraries(mudlet ws2_32)
...
endif()
```
specifies an all-lower case (library file) name - this is possibly okay
because of the way that Windows handles file names but it is probably
better to use the same thing all around.

I note that though we list both `libws2_32` and `liboleut32` in the QMake
project files it seems that only the former needs (?) to be mentioned in
the CMake case - limited experimentation I've done on my setup suggests
that the latter is linked in automagically without being explictly stated!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>