Infrastructure: tidy up naming of SLOT methods and their usage - Part 3
When Qt's slot/signal system is used to invoke a method there is some
overhead - so it makes sense to ensure developers can spot all such
methods (functions). We have tried to do this with a `slot_` prefix to the
methods we create but it has not been applied uniformly. This PR (and
one or more to follow) is intended to help with this by more rigorously
doing so - the names changed herein should all follow a
`slot_`*camelCaseMethodName* style. In addition some methods were detected
that are not currently used or which are not actually used via the
signal/slot system, these have been commented out or have had the prefix
removed and the declaration in the relevant header file moved as
appropriate.

For reference the changes made are:
* `dlgModuleManager::slot_help_module()`
                                  ==> `dlgModuleManager::slot_helpModule()`
* `dlgModuleManager::slot_install_module()`
                               ==> `dlgModuleManager::slot_installModule()`
* `dlgModuleManager::slot_module_changed(...)`
                            ==> `dlgModuleManager::slot_moduleChanged(...)`
* `dlgModuleManager::slot_module_clicked(...)`
                            ==> `dlgModuleManager::slot_moduleClicked(...)`
* `dlgModuleManager::slot_uninstall_module()`
                             ==> `dlgModuleManager::slot_uninstallModule()`
* `dlgNotepad::slot_text_written()` ==> `dlgNotepad::slot_textWritten()`
* `dlgPackageExporter::slot_import_icon()`
                                ==> `dlgPackageExporter::slot_importIcon()`
* `dlgPackageExporter::slot_export_package()`
                             ==> `dlgPackageExporter::slot_exportPackage()`
* `dlgPackageManager::slot_install_package()`
                             ==> `dlgPackageManager::slot_installPackage()`
* `dlgPackageManager::slot_item_clicked(...)
                           ` ==> `dlgPackageManager::slot_itemClicked(...)`
* `dlgPackageManager::slot_toggle_remove_button()`
                         ==> `dlgPackageManager::slot_toggleRemoveButton()`
* `dlgPackageManager::slot_remove_packages()`
                             ==> `dlgPackageManager::slot_removePackages()`

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>