Revise: tweak the map auto-save code to debug faulty invocations
Also add debugging code to track where map autosaves are being triggered;
this revealed that one place in `(void) T2DMap::mouseMoveEvent(...)` was
bogus and causing an unnecessary auto-save. It also seemed useful to
disable initiating an auto-save whilst the profile was being loaded and all
the scripts were being evaluated (for correctness amongst other reasons!)

As the `#define` needed to control the debugging code needs to be in more
than one file its definition has been placed in the project files for each
build system rather than being left at the top of a `.h` or `.cpp` file.
This lead me to promulgate #6314 to do the same for all the other bits of
debugging code that has (mainly, if not entirely by myself) been placed
around the projects source code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>