Improve: make text wrapping quicker (#6277)
Avoided duplicate call to horizontalAdvance, basically skip the first call since it was called before the function is called

<!-- Keep the title short & concise so anyone non-technical can understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
horizontalAdvance is an expensive call, current implementation in binarySearchHorizontalAdvance has a duplicate call that happen when the function is called for the first time for the line being wrapped.  This PR avoided that. 

#### Motivation for adding to Mudlet
This would have reduce the horizontalAdvance call for line that needs to be wrapped as follow:
Wrapped 1 time : 50%
Wrapped 2 times : 33.3%
Wrapped 3 times : 25%
Wrapped 4 times : 20%
Wrapped 5 times : 16.6%
Wrapped 6 times : 14.28%
Wrapped 7 times : 12.5%
Basically, 1/(wrapped times + 1).

#### Other info (issues closed, discussion etc)