Fix: allow non-integer movement of Map Labels in 2D mapper
We were rounding the coordinates which was forcing the movements to only
be to integer values for the X and Y coordinates. It might be worth
revisiting the code in this area in the future so as to provide a "snap-to-
grid" capability (but with a smaller step size than the unit space of the
room coordinates that this defect was causing!)

Also optimise some bits of code by only doing them if they are needed,
i.e. by placing them inside/after some `if (...) {...}`s that might cause
their results to be skipped/unused.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>