Add: accessible code editor (edbee) (#6283)
#### Brief overview of PR changes/additions
Add an accessible code editor so visually impaired players are able to create scripts within Mudlet.
#### Motivation for adding to Mudlet
Better a11y support.
#### Other info (issues closed, discussion etc)
This will make both the main display and the code editor accessible which are two fantastic milestones. The tab order of elements is still not right, but that is out of scope here.