Fix: prevent crash when TToolBar is turned into a_TEasyButtonBar
It seems that although we added new `TToolBar`s and `TEasyButtonBar`s to
a pair of `std::list<QPointer<T>>` containers we were not removing them
from those containers when they were (or should) have been deleted. We were
also not removing them from their owner `TAction`s when changing from one
to the other.

Also:
* change a couple of `QString::size() > 0` to `!QString::isEmpty()` as that
better represents the spirit of the test
* remove some `(bool)` local variables as a pointer assigned a value in the
same process that would be setting that boolean to `true` would also be
setting the pointer to an equally testable non-`nullptr` value.
* rename a couple of methods called `getXxxxList()` that return a
`std::list<QPointer<Xxxx>>` to `regenerateXxxxs()` that are `void` methods.
This is because the return values are not used (and the data that would
have been returned is actually stored in members of the `ActionUnit` class;
furthermore this sort of miss-naming things is something that the additions
of #6246 are trying to eliminate.

This should close #6293.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>