Fix: correct a bug exposed by prior commit
Now that we are properly deleting `TToolBar` instances their removals now
causes crashes in the code that records where `QDockWidget` derived
class instances are positioned. It probably also explains a comment I made
in #4040!

To improve the situation I have converted `Host::mToolbarLayoutChanges`
from a `QList<TToolBar*>` to a `QList<QPointer<TToolBar>>` which will help
to identify if the pointed to items have been deleted.

There were also a couple of places in the modified method where I was
seeing warnings about C++11 "Range loop detach"	potentially happening on
non-`const` Qt containers in range-for loops - which can be fixed
by wrapping the container in `qAsConst(`...`)`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>