Infrastructure: change badly named ActionUnit::getXxxxBarList() methods
There are a couple of methods called `getXxxxList()` that return a
`std::list<QPointer<Xxxx>>` which are changed here to `regenerateXxxxs()`
that are `void` methods. This is because the return values are not used
(and the data that would have been returned is actually stored in members
of the `ActionUnit` class; furthermore this sort of miss-naming things is
something that the additions of #6246 are trying to eliminate.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>