Add: support for OpenBSD operating system
Tweak the built-in Hunspell spelling support system to accomodate the
dictionaries that OpenBSD provides (unfortunately American English is NOT
one of them so we cannot use that one to locate the OS provided ones, so
instead, for that OS, look for the British one instead)...

Adjust the search for the Lua executable in
`./translations/translated/updateqm.pri`
so that for the non-Windows case it is silent when a particular executable
is NOT found (as OpenBSD at least will moan, to stderr, otherwise on the
first tries filename - which is a little confusing when the second try
succeeds!)

Rename `./src/AnnouncerLinux.cpp` to `./src/AnnouncerUnix.cpp` on the basis
that it can be used by more than just the GNU/Linux OS/

Add names for the OpenBSD specific Hunspell dictionaries so that the
"Input line" Tab of the Mudlet preferences displays meaningful names for
the found "system" (or in the OpenBSD case those packaged for the Mozilla
email and web-browser) dictionaries.

At present the CMake project build system does not seem to work on OpenBSD,
it seems that the Qt library finding stuff does not correctly find the
required Qt5 modules for one or more of the third party components.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>