Enhance: add sysBufferShrinkEvent for when a console's buffer is trimmed
This ought to help if any line number indexes are being used to access
the buffer underlying a console as then they can be revised downwards to
match the number of lines removed from the buffer/console concerned.

This commit also improves the UI for the setConsoleBufferSize(...) Lua API
function to bring it up to our current design style. It will now return
true on success or nil + an error message otherwise.

This will close #6366...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>