Improve: allow isActive(...) to use an ID number
With this PR passing an actual number as the first argument will use it as
an ID number (provided it rounds to a zero or greater integer) and return
a `1` or `0` if the corresponding Mudlet item is "active" or not (though
like the unmodified one that returns a count of items of the selected type
that are active it does not consider whether all the item's "ancestors" are
active as well before including an item in the count).

This can be used to identify the active status of temporary items (which
are only identified by an ID number) - provided of course that they are
around long enough to be checked. If the item does NOT exist then a `nil`
and an error message will be returned instead.

This ought to help with the sort of issue that #6397 raises though a script
using this revised function would, I think, be needed to do what is wanted
there.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>