Fix: return room listing widget to be within mapper
This will close issue #6369 which was brought about by my PR #5937 - but
how isn't obvious to me. Somehow before that PR the `(QTreeWidget)
T2DMap::mMulitSelectionListWidget` was automagically getting the idea of
what its parent widget was - but afterwards it wasn't - so it no longer
had a widget to position itself relative to when `QWidget::move(int, int)`
was called on it and instead moved relative to (on my GNU/Linux box) the
top left corner of the desktop. Given the behaviour that @Kebap noticed in
the original issue it seems that the Windows PC he noted it on used a
different reference position. Nevertheless I feel confident that now that
this PR explicitly tells the widget who its "daddy" (well parent) is, it
will properly position itself back to where it should be!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>